Combinatorial Generation Engine/requirements.txt
numpy == 1.21.6
Pillow == 9.1.0
psd_tools == 1.9.22
